Partial progress on linting unit test files
