+++ instruments/instruments/tests/test_util_fns.py
@@ -1,26 +1,8 @@
-#!/usr/bin/python
+#!/usr/bin/env python
 # -*- coding: utf-8 -*-
-#
-# test_util_fns.py: Tests various utility functions.
-#
-# © 2013-2015 Steven Casagrande (scasagrande@galvant.ca).
-#
-# This file is a part of the InstrumentKit project.
-# Licensed under the AGPL version 3.
-#
-# This program is free software: you can redistribute it and/or modify
-# it under the terms of the GNU Affero General Public License as published by
-# the Free Software Foundation, either version 3 of the License, or
-# (at your option) any later version.
-#
-# This program is distributed in the hope that it will be useful,
-# but WITHOUT ANY WARRANTY; without even the implied warranty of
-# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
-# GNU Affero General Public License for more details.
-#
-# You should have received a copy of the GNU Affero General Public License
-# along with this program. If not, see <http://www.gnu.org/licenses/>.
-#
+"""
+Module containing tests for util_fns.py
+"""
 
 # IMPORTS ####################################################################
 
@@ -39,6 +21,8 @@ from instruments.util_fns import (
 from enum import Enum
 
 # TEST CASES #################################################################
+
+# pylint: disable=protected-access,missing-docstring
 
 
 def test_ProxyList_basics():
@@ -123,7 +107,7 @@ def test_ProxyList_invalid_idx_enum():
 
     proxy_list = ProxyList(parent, ProxyChild, MockEnum)
 
-    proxy_list['c']  # Should raise IndexError
+    _ = proxy_list['c']  # Should raise IndexError
 
 
 @raises(IndexError)
@@ -138,7 +122,7 @@ def test_ProxyList_invalid_idx():
 
     proxy_list = ProxyList(parent, ProxyChild, range(5))
 
-    proxy_list[10]  # Should raise IndexError
+    _ = proxy_list[10]  # Should raise IndexError
 
 
 def test_assume_units_correct():
